requirements.txt
Flask
tensorflow
Pillow
numpy
scikit-learn
requests
gunicorn
psutil
waitress